refactor: move Cursor source files to Editor directory
