+++ tests/CMakeLists.txt
@@ -19,8 +19,8 @@ target_include_directories(tests
 
 target_sources(tests
     PRIVATE
-        "${PROJECT_SOURCE_DIR}/src/Cursor/Cursor.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Cursor/Cursor.cpp"
+        "${PROJECT_SOURCE_DIR}/src/Editor/Cursor.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Editor/Cursor.cpp"
         Cursor.test.cpp
 
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.hpp"
